[CT-3] 리뷰 반영 - Module 통폐합
