--- data/src/main/java/com/yessorae/data/di/CommonNetworkModule.kt
@@ -1,29 +1,0 @@
-package com.yessorae.data.di
-
-import com.yessorae.data.BuildConfig
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-import okhttp3.Call
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-
-@Module
-@InstallIn(SingletonComponent::class)
-object CommonNetworkModule {
-    @Provides
-    @Singleton
-    fun provideOkHttpCallFactory(): Call.Factory =
-        OkHttpClient.Builder()
-            .addInterceptor(
-                HttpLoggingInterceptor()
-                    .apply {
-                        if (BuildConfig.DEBUG) {
-                            setLevel(HttpLoggingInterceptor.Level.BODY)
-                        }
-                    }
-            )
-            .build()
-}
