+++ data/src/main/java/com/yessorae/data/di/DatabaseModule.kt
@@ -3,6 +3,10 @@ package com.yessorae.data.di
 import android.content.Context
 import androidx.room.Room
 import com.yessorae.data.source.local.database.ChartTrainerDatabase
+import com.yessorae.data.source.local.database.dao.ChartDao
+import com.yessorae.data.source.local.database.dao.ChartGameDao
+import com.yessorae.data.source.local.database.dao.TickDao
+import com.yessorae.data.source.local.database.dao.TradeDao
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -21,4 +25,17 @@ object DatabaseModule {
             klass = ChartTrainerDatabase::class.java,
             name = ChartTrainerDatabase.NAME
         ).build()
+
+    @Provides
+    fun providesChartDao(database: ChartTrainerDatabase): ChartDao = database.getChartDao()
+
+    @Provides
+    fun provideChartGameDao(database: ChartTrainerDatabase): ChartGameDao =
+        database.getChartGameDao()
+
+    @Provides
+    fun provideTickDao(database: ChartTrainerDatabase): TickDao = database.getTickDao()
+
+    @Provides
+    fun provideTradeDao(database: ChartTrainerDatabase): TradeDao = database.getTradeDao()
 }
