+++ data/src/main/java/com/yessorae/data/di/NetworkModule.kt
@@ -1,5 +1,6 @@
 package com.yessorae.data.di
 
+import com.yessorae.data.BuildConfig
 import com.yessorae.data.source.network.polygon.api.PolygonChartApi
 import com.yessorae.data.source.network.polygon.common.PolygonConstant
 import dagger.Module
@@ -8,15 +9,31 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 import okhttp3.Call
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 @Module
 @InstallIn(SingletonComponent::class)
-object PolygonNetworkModule {
+object NetworkModule {
     @Provides
     @Singleton
-    fun provideRetrofit(okhttpCallFactory: Call.Factory): Retrofit =
+    fun provideOkHttpCallFactory(): Call.Factory =
+        OkHttpClient.Builder()
+            .addInterceptor(
+                HttpLoggingInterceptor()
+                    .apply {
+                        if (BuildConfig.DEBUG) {
+                            setLevel(HttpLoggingInterceptor.Level.BODY)
+                        }
+                    }
+            )
+            .build()
+
+    @Provides
+    @Singleton
+    fun providePolygonRetrofit(okhttpCallFactory: Call.Factory): Retrofit =
         Retrofit.Builder()
             .baseUrl(PolygonConstant.BASE_URL)
             .callFactory(okhttpCallFactory)
@@ -25,6 +42,6 @@ object PolygonNetworkModule {
 
     @Provides
     @Singleton
-    fun provideChartApi(retrofit: Retrofit): PolygonChartApi =
+    fun providePolygonChartApi(retrofit: Retrofit): PolygonChartApi =
         retrofit.create(PolygonChartApi::class.java)
 }
